+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 For a list of breaking changes, check [here](#breaking-changes).
 
+## Unreleased
+
+- [#1794](https://github.com/clj-kondo/clj-kondo/issues/1794): Add a linter for line length
+
 ## 2022.09.08
 
 - `:config-in-call` - see [docs](https://github.com/clj-kondo/clj-kondo/blob/master/doc/config.md#config-in-call)